Download and store set and printing information in db.
Sets are MtG sets. Pretty straightforward. They go in the table `set`.
You'll always want backticks on that because SET is an SQL keyword but
they are so very definitely called sets that I used it anyway.

Printings are individual printings of cards. Effectively card+set.

I moved the properties that are not populated in AllCards.json (on a 'card')
but are populated in AllSets.json (on a 'printing') to printing.

I changed the teardown of the database from a bunch of DROP statements to
a simple delete of the file. That way you don't have to remember to add
a table to the list. That said if we ever want persistent data we'll either
have to use another sqlite db or reverse this decision.

Schema version bumped (a lot!)

I meddled with unzipping to make it generic as we now do it for two
different files.

We can now add search on the new fields like rarity and artist and fallback
images with multiverseid can now be made to work. That's upcoming.