Refactor bot now that we have command.
Various things here to appease the linter and hopefully makes things nicer
too.

    1. No more STATE. Instead, bot is an object and holds its own state.

    2. Most helper functions used by commands moved to command. Some usage
       lingers in bot which is probably bad. Might clean this up later.

    3. All commands now have the bot passed in as a variable. We can maybe
       simplify to only exposing client but for now a few other things are
       accessed.

    4. run.py got a little more complicated because I can't work out how to
       annotate Bot instance methods with @client.event.

One thing that has become wildly apparent during this is that it should be
possible to access Oracle without having to instantiate it and going through
a bunch of checking either by having a singleton or making it stateless.
I'll follow up on that.