Split card into card (physical card) and face.
In most cases these have a 1:1 relationship. But meld, split, flip and double
faced cards have a 2:1 or 3:1 relationship between face and card.

Also share the bulk of the SELECT for oracle.search and search.search so they
don't get out of whack about how to combine mulitple faces into a single
card.