Fetch legal_cards.txt as utf-8 so we interpret Seance and others correctly
There is an ongoing issue with Iff-Biff Efreet but other accents should now
be handled correctly.

Note that the RFC says anything served as text/plain without
Content-type/character encoding in the header should be treated as US-ASCII
so that's why we need to be explicit.

I'm sure latin-1 worked before but perhaps it just worked better than
US-ASCII and always messed up on accents. Or maybe there was a change on the
server. Regardless, it works now.