Clean up database usage a little.
Moving to apsw gave us proper transactions under our own control so we no longer
need to reach into the guts of database quite so much.

Still not convinced about the API database is offering to the world here but
this is at least better.

Changed self.database to self.cursor to make it clearer what it is, too.