Now a deck knows if it is pd legal or not.
I'm introducing some potentially horrendous performance and some "you have
to do X before you do Y" to decks here but I'm conscious of it and I'm
coming back to tidy up later. I'm getting a stake in the ground for what
we want before I start making it perfect. Definitely erring on the side
of messy for now.