Rename cards() to all_cards() to avoid conflict with cards property.
We take actions in View based on whether a view has a cards property and some
views pass unknown attributes onto an object they contain. We don't want
a view that wraps a Deck to respond to the request for the cards property
with a function. That won't work. (It would in ruby!)