Fix issues with timezones.
Our test for timestamp of 0 was out by an hour when the system clock was set to
Europe/London. This was because the timezone conversion is naive based on
current rules and the UK had 'British Standard Time' (GMT+0100) throughout 1970.

Move to a more sophisticated third party lib based on the near-definitize
Olson tz database.