Use date_updated for created_date in tappedout scraper. Fix login.
I switched login to https and it started working so I switched everything.

Using date_updated as a created date is a bit weak but it's the best they give
us. We don't actually do updates but as/when we start doing that we can use it
to update updated_date in subsequent updates.

Fixes #419.