Some simple response caching.
This makes subsequent request for the same path+querystring very fast.

But it does nothing for the first request.

Fairly conservative with times for now - we can maybe cache more aggressively.

A client hard-refresh will always be respected.

Not caching some things either because they are inexpensive, binary objects,
or need to be very up-to-date.