Remove complex descender skip in favor of CSS version; mark external links.
text-decoration-skip doesn't have universal support yet but it works in
latest Chrome and it's simpler and allows us to mark external links without
a silly underline under the :after.

For now we add the external marker with js. If the flash of content being
styled seems annoying we can add it in the html or process in the view
before sending to the browser.