Do section top spacing as padding not margin to avoid ugly gray line.
If an unfloated section is the first element in main using margin-top will cause
the body background (gray) to show and not white.

The body background is gray so that scroll bounce shows the right color for the
header and the footer.

section elements are not floated in narrower views so that tables within them
can be given overflow-x to scroll-in-place. Floated elements are out-of-flow
and thus cannot be given that treatment.

In other news, CSS is a bit too complicated.