Sort formats in legality.
Block legality will still be in a random order for cards legal in more than
one block but otherwise they will be in a sensible order.

I pushed this into the view for individual decks and cards to avoid doing
unnecessary work in prepare_card and prepare_deck. If this is a true policy
change on that then we should probably move more stuff out of prepare_card
and prepare_deck and into the detail page view.

Fixes #558.