Refine right-scroll section changes.
If we give this an overflow-x we get funky scrollbars on Firefox/Windows and
we clip the right-margin of tables with marginalia.

Just the max-width that we added at the same time is enough to make the
overflow-x of the contained tables do the right thing, or so it seems.

So this is theoretically the best of both worlds.

Fixes #3963.
Fixes #3954.