Remove all of our search code in favor of scryfall.
This is sad, and this code was great, but maintenance cost and usability
favors pushing this all to scryfall. !rhinos is quite a bit slower and you
can't do is:gainland but url now links to the same search and all scryfall
syntax is supported. People were mostly using !scry anyway.

Fixes #4554.
Fixes #4243.
Fixes #4153.