Season-ize (most of) person detail page.
One unhappy consequence of these changes is that if you try to visit the
person page of someone for a season they did not play in you will get a 404.
I'll fix that up in a future commit.

Cards and unique cards played are not currently seasonized but achievements,
head to head and decks are.

More to follow in the great seasonizing revolution.

(Note: to seasonize is to make a page show a season-specific view instead of
all time by default and allow changing between seasons and indeed to All Time).