Prep linter to cope with logsite.
We add alembic and flask_sqlalchemy to ignored-modules because they do magic.
We also pass these on the commandline for local tests even when the module is
installed because of the magic. We're not just ignoring them for Codacy
(which doesn't install our dependencies).

We add 'query' to generated-members (unfortunately broad but this is how
pylint works) to avoid the generated query method on our SQLAlchemy stuff
triggering a no-member error.

Exclude the whole logsite_migrations directory on the basis that we don't
want to lint generated code. This may be an issue with codacy as it isn't
running our make lint command but rather doing its own thing but let's see.