Migrate Card to shared.models.card from magic.card.
Models that are basically dumb data objects and don't need any support from
other code can be shared across all of the monolith and imported (for typing)
without worrying about crossing over into layers they should not be.

Begin the migration with Card which is so widely used it already lives in
magic and not in decksite and is imported all over the place.

I think it makes sense for Printing, Archetype, Deck and others to follow but
that can be evaluated separately.