This is the worst-performing query on the site and frequently takes >10s.
This is very much a nice-to-have feature so I'm going to remove it for now.

It actually makes no sense in that if your first use is any good someone else
will use the card so this becomes essentially your "hall of shame" of mistakes
you made! It is interesting though so I'll try to find a replacement or generate
it offline or speed up the query.