A better if not perfect mobile menu.
Everything visible from the off as (sort of) recommended in
https://www.nngroup.com/articles/hamburger-menus/
in the Time Magazine example.

To avoid it being completely un-obvious that navigation is complete
(particularly if it is fast), scroll you to page content after navigation.
I'm not quite sure how annoying/clear this last feature is but let's try
it.

Fixes #2057.