Refacor achievements.
1. Move decksite.achievements to decksite.data.achievements.

This module is mostly concerned with loading data from the database so it
belongs in data.

2. Change what is emitted from decksite.data.achievements from an entity to a
struct.

What we pass in/out of data should be as dumb a data-holding object as possible
(a Container) rather an an entity (object with logic/behavior). See Clean
Architecture.

3. Load achievement leaderboards into self.leaderboards in the view to take
advantage of decksite.views.view.prepare_leaderboards and thus remove some
code duplication.

The general contract in view is that if you have something that used in multiple
pages you can put it in the right self.{property} and view will massage it into
the right state. Applies to cards, decks, people, etc. too.

4. Minor refactors around names/imports for consistency.

Hopefully I didn't break anything :D