Screen for admins to add/view Gatherling aliases. Plumbing for same.
I also changed the SQL for the person_alias table (it hasn't gone live yet)
to better reflect what we need (we want to intercept future uses of these
alternate accounts and insert the decks under the right username, not recreate
these deleted incorrect entries in people so we have to store a string not an
id as there will be no id).

See #3880.