Create explicit stacking contexts for top-level page elements.
Initially I did this to fix #5837 (because filter:grayscale(50%) on <header>
was inadvertently creating a stacking context and the upshot was that search
results were being hidden behind the menu). But then I changed how the
grayscale filter was being applied to the header background image (because
we only want to grayscale the background image, not the whole of the header
including text, language chooser submenu, etc.) But I'm still committing
this because z-index is a tricky beast and this is a sensible approach to
containing it.