Allow scryfall-like searching of rotation cards.
This is potentially dangerous as we now allow ridiculously lengthy searches
of Scryfall but a combination of redis caching and sleeping for 100ms after
each search should limit the problems to our side of things.

Because we store a list of dicts in redis I had to despecify the return value
of redis.get_list.