Make spoiler background task and spoiler bot command use standard card inserts.
Database setup got updated to be a handful of queries total instead of a
handful of queries per card. But spoiler command and background task on the bot
continued to use the old code. There is/was also a race condition there
resulting in duplicate entries.

Here I force all card inserts to go through the same insert_cards function and
remove a lot of unused code. The insert_meld_faces stuff is now the only thing
not in the giant inserts in insert_cards. I've opened a ticket to clean that
up so we can delete yet more code.

I'm not sure I've got rid of the race condition here because the background
spoiler season task could conflict with a manual call to spolier and other
things. And it's not great to recache and reindex on user time while the bot
remains unresponsive (although that does prevent race conditions). I've opened
tickets to maybe improve these things.

Fixes #6317.
Fixes #6308.
Fixes #6272.