Add regular rotation messages from the bot during rotation season.
This meant pushing a lot of what was happening in decksite.views.rotation
into magic.rotation.

Unfortunately it is not possible to push interestingness into magic.rotation
because it needs knowledge only decksite has.

So the view still decorates the cards with interestingness after the
magic.rotation stuff.

And the bot doesn't know how interesting cards are and just lists an arbitrary
four.

It would be better to expose interestingness to the bot somehow but that will
be a future improvement.