Remove special treatment of overflow in sections/highlights.
As it currently stands this clips the loading spinner and part of the L of
Loading. If we just remove the overflow-x: auto then tables like /rotation/
that don't get footable behavior overflow into the right margin. So we remove
overflow: visible on narrow view also. I'm not sure what this was showing
before but all interestingness remains visible so I think this is ok.