Remove footable responsive tables in favor of sideways-scroll.
Fixes #5583.

Integrating react loading, latest jquery and footable just isn't going to work.

As footable is the main problem, brings dubious benefit, and affects performance I'm going to remove it from the app.

The consequence of this is that data tables will horizontal-scroll instead of having that "+" symbol on the right that allows you to see the rest of the data.

When everything uses react maybe we'll revisit doing something hand-rolled in the react tables.