A new mobile menu.
Far from perfect but I think (?) better than what we had.

The submenu in particular eats a ton of screen real estate and might
be better as a hamburger menu.

As a result we get to remove the scroll to content stuff that jumped you
beyond the huge inline menu from before. This means we no longer break
the back button on mobile.

Fixes #6271.