Say '6 days' instead of '1 week' for 6 days but keep all the desirable behavior.
We introduce here the concept of a "rounding threshold" and each the machine
that you never round up weeks, round up days after 7, hours after 24, minutes
after 45 and seconds after 31.

The rest is extra tests for this new behavior and fixes in the existing
behavior prompted by issues I found trying to change this behavior.

Fixes #7060.