Remove background_task_spoiler_season
Scryfall no longer supports the API endpoint we used for this. New cards
will come in via the spoiler command and restarts/updates only.

We might want to look at ressurrecting this in some way but it'll have to
use the bulk data endpoint (aka be a scheduled update, basically).