Remove character encoding from json.loads - it has been ignored since Python 3.1
And since 3.8 it emits a deprecation warning about being removed in Python 3.10

We DO use character encoding in fetch because it speeds things up for reasons.
I guess we should also be doing that in async? But it's not clear how, we have
no use case right now and even if it's slow at least it is async.