Have the bot clear the redis rotation cache before emitting hype message.
The semantics of decksite are quite simple - have a long-lived cache but remove
it when a run completes.

Because the bot does not use the same redis db as decksite and the rotation
script this won't work for the bot. Instead clear the cache immediately
before generating the hourly message. Cache will still be used for `!hype`
command but maybe it makes sense to clear it there too.