Make Snow and Colorless symbol sort ordering stable, fix issues with sort_score
Because we were ignoring Snow and Colorless entirely when deciding on an order
of mana symbols for display we were storing inconsistent values in
deck_cache.colors.

Separately our sort_score was not actually sorting the way we want colors
sorted using the display order as input and doing things like making
['W', 'G'] and ['G', 'B'] get the same score. Switch to a simpler system
that uses 2-to-the-power-of-position like unix file permissions and similar.
But put even more weight on the number of colors so that monocolor decks always
sort before 2 color decks and so on.

Fixes #9988.