Performance improvement when generating _card_person_stats
This shows a mild performance improvmnt on local but a radical improvment on prod. From basically never completing (50 minutes and not finished) to just over three minutes.

Looking at the EXPLAIN output this removes "Using filesort" because the results are no longer being sorted in the order implied by the GROUP BY.

We should probably apply this to all preaggregation.