Don't prepare decks on card detail page
On a page like /seasons/all/cards/Dark Ritual/ this can cause prepare.prepare_deck to be called for several seconds and we no longer use the prepared decks for anything. We used to display them in a table but we load them via XHR these days.

Ideally we shouldn't be loading 14,000 decks at all but as an immediate improvement let's get rid of a bunch of busy work that brings no benefit.

See #11230.