Make it explicit that smoke test needs a test db
The new test collector in pytest 8 was outputting the tests in a different
order and because smoke_test was now running before gatherling_test it was
failing. I guess it was relying on the with_test_db in gatherling_test to work
at all before.

This is more work as we're now recreating decksite_test and applying all the
patches to it another time but as it's for functional tests and relatively
quick I think it's ok?