Allow upgrade to pytest 8 by fixing some funkiness with db and with_test_db
This is kind of complicated.

It used to be that something about how the pytest test collector worked meant that gatherling_test (which uses the with_test_db wrapper to spin up an empty copy of the decksite db before running the test) and decksite's smoke_test were ordered in such a way that when smoke_test (which doens't use with_test_db) ran there was a sane database available and it was able to load the home page and /cards/Unsummon, etc.

Something about the changes to pytest collector in v8 (they were extensive) caused this to be reordered in such a way that the smoke_test was failing unable to find the tables it expected in the decksite db it was looking at. It would call 'db()' which, if it doesn't find the db it's expecting, creates the db if missing, but does not apply any migrations. It would then fail to find the `source` table (from migration 1) and fail the test.

I haven't fixed this in the spiffiest possible way. All I've done is change two things:

1. A database is now stored under its name inside the context. So instead of having one (g|ctx).database we instead make (g|ctx).decksite and (during with_test_db tests at least) (g|ctx).decksite_test. So you'll never ask for "the database" in some code and get one that is not the one currently configured.
2. I have changed with_test_db to put configuration.CONFIG['decksite_database'] back the way it found it after running the test in question. Before this was having the side effect of having every test run subsequently look at decksite_test instead of decksite.

This makes the tests pass under pytest 8. What it doesn't do it is mark any tests that touch the database as with_test_db. This means that the local setup at the time the tests run probably affect whether they pass or not. I guess it would be best to mark all tests that require a decksite db with with_test_db. It seems wrong that we would have tests that require a decksite db but don't have this annotation. I have logged #12093 to maybe get us to that joyful land some time in the future. I'm not sure at this time what the ramifications of that are.

But this at least returns everything to working order and gets us to pytest 8. No crimes greater than those previously being committed are added.