Reinstate editing a deck's archetype from the deck detail page
This was removed thinking that it didn't hurt anything to improve performance
and it does in fact make the page take 0.07s instead of 0.6s on local for me
but the performance hit is now only for mods and not so bad to worry about I
think. Having the ability is very useful.