Don't show all legal seasons in deck table, show the deck's season
The main reason for doing this is so that the pagination arrows are always
in the same place rather than "jumping" if a deck would be legal in more
than one season.

It's honestly probably more interesting to see what season a deck was played
in than all the seasons it would have been legal, too.

Also we don't re-cache this value very often so what we used to show was not
even accurate a lot of the time.

Fixes #77690.