Standardize trailing slash on paths
We have a lot of inconsistency. This commit brings consistency.

The rules are now enshrined in a functional test. They are as follows:

- If you look like a file (/foo/bar.txt) then you don't have a traililng slash
- If you are part of the API you exist with both the trailing slash and without (no redirect)
- If you a "normal" web page then you exist only with a trailing slash, Flask will redirect requests without the trailing slash to the canonical URL

It was too difficult to enforce a standard for the API from where we started so rather than have a mish-mash we will always support both.

A couple of things are grandfathered into having no trailing slash (oembed and logite's match export) just in case anything is relying on those, but the rules are cast iron and enforced going forward!