Use CSS special properties (variables) for colors
While I was in there I found a little bit of dead CSS and also a tiny issue
from when page size used to be in the pagination section of the live tables
but no longer is. I removed the dead code and fixed the issue.

I also found we were using #ccc as a third gray. I removed it in favor of one
of our other two grays in each case. Sometimes this was a decent impromvement
like the mobile menu background. The only worse case is achievements "more info"
but it didn't seem worth keeping a whole color for that.