Fix issues with playability where a card was not played when legal
"The Kapsho Kitefins issue". Kapsho Kitefins was (until now) rated as one of
the top 1000 or so PD cards (of the 21,205 that have ever been legal). This
was because it was played once, in season 1, when there were only 150ish
decks. The playability code was saying "you were playing in 1/150 of all the
decks in the relevant seasons, you're pretty good". When it should have been
saying "you were played once in all 32 seasons of you being legal in
1/200,000 decks in the relevant seasons, you totally suck!"

deck, deck_card and deck_cache (heretofore the source of truth for all this)
don't know anything about what cards were legal when except inasmuch as they
know what decks contained what cards and what seasons they are from. They
are unable to differentiate between "not legal", "legal but never played" or
even "hadn't been printed yet". To get around this I introduce a
_legal_cards cache table that holds {season_id,card} pairs as rows. These
can then be used in the creation of _playability which needs to know (for
use as the denominator) the number of decks that a card could have been in
(aka the total of all the decks in the seasons it was legal, NOT the seasons
in which is was played, which can be a subset of the seasons it was legal).

I create _legal_cards by calling search.search('f:pd{season_id}') which is
honestly pretty dubious? But cards db is the source of truth for this and
that's how decksite talks to cards db? So it's ok? It's better than a cross-
db query at least. I think.

This has the incidental effect of routing around issues where decks are
assigned to the wrong season. The playability score of Necropotence,
Braistorm and Demonic Consultation are all currently depressed by the fact
that a deck or two containing these cards is currently assigned to the wrong
season_id in deck_cache. We still need to fix that but this will mitigate
the issue (because the base table for the query is now _legal_cards which is
100% accurate).

Fixes #9453.