Highlight cards that are/were newly-legal on /rotation/changes
This fixes a little regression we had when switching to the new output
format at /rotation/changes where we lost that info.

I'm not concerned about the size of the output of /api/cards2 - it's a few
KB gzipped and all the 1-400ms we spend here is server time not transfer
time. The extra work being done per-card is also miniscule and in the
reactified world should be done a maximum of 100 times.

Fixes #12417.