Get total results when fetching slice of results for cards
This is way less error prone than having two queries that do essentially the
same thing and must be updated when each other is updated. We could share the
code but this is even easier.

The performance penalty on /seasons/all/cards seems to be negligible with the
total query frequently beating the old query in my testing on prod.

See #12412.