Give the site a bit more depth
This mainly comes in the form of giving things that are "above" other thing
and can occlude them a drop shadow.

I also took the opportunity to massively simpify our z-index and bring it in
line with Material Design, from whence I also nicked the drop shadows.

I had to mess a little with some existing stuff so that things that were
before a higher z-index are now further down in the HTML so they are still
clickable. I didn't manage to make this work with the archetypes to sort
badge so that's TODO.

The end result is mostly that:

  1. The new season picker has some more padding and a drop shadow making
     it more obviously "above" the page.
  2. The overlay submenu (desktop only) has a drop shadow.
  3. The status bar has a drop shadow (only relevant on mobile when it goes
     to the top of the screen).
  4. The card image that shows on hover of a card name has a drop shadow.

I didn't manage to get a drop shadow working on text tooltips but as that's
at the highest level of elevation getting that working is also a TODO.