Remove long-failing test
Because whatsinstandard doesn't go back to S1 if we want to make these functions
work from any start date we'd need to switch to some other method. We could use
the seasons table in decksite but that would make this decksite-specific.

In general it doesn't seem worth it to make a far-reaching change on how we
think about seasons and pd500 dates and so on just to have some better tests so
I'm removing this.