Work around Chrome/iOS and Safari issues by merging all our symbol fonts
For whatever reason these browsers only respect emoji and similar from the
*first* font listed and will fallback to system fonts after that. So here
we provide one font with all the symbols we care about in from a collection
of subsetted fonts.

In order to do the merge I had to switch us to simpler versions of the CJK
font as fonttools is unable to merge the variable CJK font.

Now we just emit one @font-face called "symbols".

Add the hamburger and home icon we are using in the new menu.