Speed up /admin/archetypes
All numbers from local when there were 180 decks to sort.

(1) Add a LIMIT which speeds up the main slow query (5s) so that the total page load goes from about 7s to about 3.5s

(2) Add an INDEX which speeds the query up from 5s (without LIMIT) to 400ms (with LIMIT).

There is no pagination here, you have to refresh the page to see the next 20 decks. And it's still kinda slow. But a decent band-aid.