Restore full width of marginalia
I reduced this when I added the menu but it causes alignment issues between h2 section headings and tables with marginalia.

A three star deck will now be kind of too close to anything on the left but we can live with this for now. A new grid is coming soon, I think.