Remove not-in-rotation decorator, it's needless complication
It's ok to update the rotation cache if asked to when it's not rotation time
and that mostly won't happen anyway. Less code good.

Also this will allow me to bootstrap the _rotation table on the newer prod
box even though it's not currently rotation time now that rotation is going
to be handled there. This will reduce log spam when calling rotation pages
until next rotation in 8 weeks.