Introduce standard text padding
These are issues in the design mostly made plain by adding the
background-color-on-hover behavior to trs. The background color looks
truncated and bad.

This means that both left and right-aligned things can have a background color
without looking terrible.

Also make spacing above/below mana-bar on tables consistent because the bg
color on trs exposed that as lopsided, too.