Outline form elements when focused, rather than making them 80% opacity
In 2017 I got rid of the browser-default outline because it was ugly. I replaced
it by reduced opacity, not sure why. In 2021 I move the opacity to 80% because
the previous value was too extreme.

In general I don't think this behavior exactly makes sense and it makes the
search box look kind of stupid in dark mode when you click in it. So lets go
back to an outline, but use our highlight color.