Make highlit table rows clickable, remove highlighting where that's wrong
There are a few here that need some work in the data part before we can link
them (/admin/bans, /admin/people/notes) and others that have inline form
elements that the simple js interferes with. But most every other table row
in the site that has a sensible default link now links to that place when
clicked anywhere that doesn't explicitly link somewhere else.