Don't clear decksite rotation cache via web, be more direct
Previously this was timing out when we got to later runs. So I made it run in
a thread not on user time. Which resulted in it crashing the webserver on local.

Here we block the progress of the rotation script to do this, which is not super
pleasing. But we catch all Exception and continue so it should be pretty hard
for this to break rotation.

And this way the bot should never be announcing things the site doesn't know yet
and other slightly race condition-y stuff.