[download_scryfall_card_image] fix double-sided images when not JPG
For double-sided cards, both sides are downloaded then combined into a single
image.  The filename generation assumed the extension was `.jpg` with the result
that when it was `.png`, both images were written to the same path, and we got a
merged image of two copies of the card back.