Ignore type issues stemming from Beautiful Soup's new(ish) type hints
All these funcs now return PageElement | Tag | None. PageElement can't be
treated like a dict. So all this code needs either a bunch of casts or a bunch
of instanceof checking or some type: ignores.

Honestly it doesn't feel like this is sustainable. I think they'll find a better
way. So I just type: ignore-d for now and posted on the bug tracker:

https://bugs.launchpad.net/bugs/2099772

Hopefully at some point we can remove these. Or we may have to reluctantly do
a bunch of cast-ing or something.