Update fonts maintenance task to generate an actual woff2 file
Then point at this woff2 file in the CSS.

Now we can run the task any time, on prod.

This means adding some large font files, which I did using LFS. "Normal" git
will check these out as pointers and only once you install git-lfs and
`git lfs pull` will you get the actual fonts.

It's also the case that you'll need to run the maintenance task to generate the
font rather than getting it for free with your checkout:

    $ pipenv run python run.py maintenance fonts

However the benefit is that we will now be able to trigger a regeneration on
prod whenever a deck is added with an unusual symbol and this will no longer
be a manual process run once every few months. Hurray!