[download_scryfall_card_image] fix for dfc with version=border_crop
previously the code generating the pair of paths for dfcs assumed that `version`
would be the extension and would be contained in `filepath`.  however, `version`
can also be `border_crop`, so instead of calling a replace on `version` we
split the filepath on the extension and insert the a/b strings.