Get the real stack trace into GitHub issues
We were getting:

Traceback with variables (most recent call last):
werkzeug.exceptions.InternalServerError: 500 Internal Server Error: The server encountered an internal error and was unable to complete your request. Either the server is overloaded or there is an error in the application.
500 error at /admin/rules/

for basically every error on decksite, but not the actual real stack trace
(which IS in the error log).