Defend against multiple signups in the same second from the same player
I'm not totally sure HOW this happens, but we do see it. Seems completely
reasonable to send you to the deck of the first submission on subsequent
requests. If you submit two DIFFERENT league decks in the same second this
will kind of be a lie but really at that point it's on you. How are you even
doing that? :)

Script to test this at https://gist.github.com/bakert/e0408594f211c117fc7d6edb634ca778
should we need it.