Use the underlying original exception not the skinny werkzeug wrapper
This will give us real stack traces in our GitHub issues instead of just the
generic InternalServerError message.

We no longer need to redact here because we're just using format_exception
and not format_exception_with_variables. The latter gives more detail but is
so much harder to read that I'm just going with the simple version. When we get
a bug that requires the vars to debug I'll kick myself then :)